chore: move lib into helpers folder
