+++ count_ingredients.c
@@ -1,4 +1,4 @@
-#include "ingredients.h"
+#include "./helpers/ingredients.h"
 
 #define MAX_LINE_LENGTH 1024
