+++ find_biggest_repeatable_ingredients.c
@@ -1,4 +1,4 @@
-#include "ingredients.h"
+#include "./helpers/ingredients.h"
 
 typedef struct {
     char *value;
